% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-control-seurat.R
\name{qc_seurat}
\alias{qc_seurat}
\title{Quality Control and Normalization of a Seurat Object}
\usage{
qc_seurat(input)
}
\arguments{
\item{input}{is the seurat object with .rds extension}
}
\description{
This assumes that the input file is already a seurat object in an .rds format.
Other files/extensions that is not a seurat object in rds format should be converted first.
A package "write_seurat_object" is available on git
This package contains three functions: for the quality control \
of the seurat object, for the normalization of the seurat object \
after QC, and for the visualization of the seurat object during \
quality control
}
\examples{
quality_control_seurat("seurat_object.rds") # will output a file after quality filtering the output file should be used as the input for normalization
normalization_seurat("afterQC_seurat.rds")
}
